kernels: Zero-initialize intermediate output memory region

+++ sw/applications/src/utils/utils.h
@@ -3,4 +3,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 void check_layer(layer l, double* checksum);
-void dma_memset(void *ptr, int32_t value, uint32_t len);
+void dma_memset(void *ptr, uint8_t value, uint32_t len);
